// client/src/app/client/model/laptop.ts
export interface Laptop {
    laptop_id: number;
    laptop_mark: string;
    laptop_description: string;
    laptop_price: number;
    laptop_status: string;
    laptop_system: string;
    laptop_processor: string;
    laptop_ram_size: number;
    laptop_disc_type: string;
    laptop_disc_size: number;
    laptop_graphics_card: string;
    laptop_screen_diagonal: number;
    laptop_amount_available: number;
  }